[REFACTOR] #162: 알림 삭제 시 알림 유효성 검사 시행 및 에러코드 작성

+++ src/main/java/project/luckybooky/domain/notification/service/NotificationService.java
@@ -138,6 +138,10 @@ public class NotificationService {
      */
     @Transactional
     public void deleteByUser(Long userId, Long notificationId) {
+        boolean exists = notificationRepository.existsById(notificationId);
+        if (!exists) {
+            throw new BusinessException(ErrorCode.NOTIFICATION_TYPE_NOT_FOUND);
+        }
         notificationRepository.deleteByUserIdAndId(userId, notificationId);
     }
